+++ Firmware/MotorControl/axis.cpp
@@ -93,6 +93,19 @@ void Axis::set_step_dir_enabled(bool enable) {
     }
 }
 
+bool Axis::check_for_errors() {
+    // Maybe we should update this to only trigger on new errors?
+    // The danger with that is we could fail to bail on uncleared errors that still prevent
+    // correct opreation.
+
+    // For now: we treat ERROR_INVALID_STATE in idle loop special, or we could never stay
+    // in idle after this kind of error.
+    if (current_state_ == AXIS_STATE_IDLE)
+        return (error_ & ~ERROR_INVALID_STATE) == ERROR_NONE;
+    else
+        return error_ == ERROR_NONE;
+}
+
 // @brief Do axis level checks and call subcomponent do_checks
 // Returns true if everything is ok.
 bool Axis::do_checks() {
@@ -112,7 +125,7 @@ bool Axis::do_checks() {
     // sensorless_estimator_.do_checks();
     // controller_.do_checks();
 
-    return error_ == ERROR_NONE;
+    return check_for_errors();
 }
 
 // @brief Update all esitmators
@@ -120,7 +133,7 @@ bool Axis::do_updates() {
     // Sub-components should use set_error which will propegate to this error_
     encoder_.update();
     sensorless_estimator_.update();
-    return error_ == ERROR_NONE;
+    return check_for_errors();
 }
 
 float Axis::get_temp() {
@@ -159,7 +172,7 @@ bool Axis::run_sensorless_spin_up() {
     // is zeroed. So we make the setpoint the spinup target for smooth transition.
     controller_.vel_setpoint_ = config_.spin_up_target_vel;
 
-    return error_ == ERROR_NONE;
+    return check_for_errors();
 }
 
 // Note run_sensorless_control_loop and run_closed_loop_control_loop are very similar and differ only in where we get the estimate from.
@@ -178,7 +191,7 @@ bool Axis::run_sensorless_control_loop() {
         return true;
     });
     set_step_dir_enabled(false);
-    return error_ == ERROR_NONE;
+    return check_for_errors();
 }
 
 bool Axis::run_closed_loop_control_loop() {
@@ -193,7 +206,7 @@ bool Axis::run_closed_loop_control_loop() {
         return true;
     });
     set_step_dir_enabled(false);
-    return error_ == ERROR_NONE;
+    return check_for_errors();
 }
 
 bool Axis::run_idle_loop() {
@@ -203,7 +216,7 @@ bool Axis::run_idle_loop() {
     run_control_loop([this](){
         return true;
     });
-    return error_ == ERROR_NONE;
+    return check_for_errors();
 }
 
 // Infinite loop that does calibration and enters main control loop as appropriate
@@ -249,9 +262,10 @@ void Axis::run_state_machine_loop() {
                 task_chain_[pos++] = requested_state_;
                 task_chain_[pos++] = AXIS_STATE_IDLE;
             }
-            task_chain_[pos++] = AXIS_STATE_UNDEFINED;
-            // TODO: bounds checking
+            task_chain_[pos++] = AXIS_STATE_UNDEFINED; // TODO: bounds checking
             requested_state_ = AXIS_STATE_UNDEFINED;
+            // Auto-clear any invalid state error
+            error_ &= ~ERROR_INVALID_STATE;
         }
 
         // Note that current_state is a reference to task_chain_[0]