+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Unreleased Features
 Please add a note of your changes below this heading if you make a Pull Request.
 
+## Unreleased
+### Added
+* Encoder position count "homed" to zero when index is found.
+
+### Changed
+* We now enforce encoder offset calibration must happen after index is found (if using index)
+
+### Fixed
+* Once you got an axis error `ERROR_INVALID_STATE` you could never clear it
+
 # Releases
 ## [0.4.2] - 2018-08-04
 ### Added